requirements.txt
Flask==2.3.3
Flask-Cors==1.0.0
Pillow==10.0.1
numpy==1.24.4
torch==2.0.1+cpu
torchvision==0.15.2+cpu
matplotlib==3.7.2
seaborn==0.12.2
ultralytics==8.0.196
opencv-python-headless==4.8.1.78
gunicorn==21.2.0